azure_monitor/python_logger_opencensus_azure/monitoring_with_singleton/requirements.txt
opencensus-ext-azure
opencensus-ext-logging
opencensus-ext-flask
opencensus-ext-requests
mlflow